default to using POST requests
